# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Blaze1518_sinhnhatf168_internal_errors.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  github_com_Blaze1518_sinhnhatf168_internal_errors.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      path:
        type: string
      request_id:
        type: string
      retry_after:
        type: integer
      timestamp:
        type: string
    type: object
  github_com_Blaze1518_sinhnhatf168_internal_errors.Links:
    properties:
      first:
        type: string
      last:
        type: string
      next:
        type: string
      prev:
        type: string
      self:
        type: string
    type: object
  github_com_Blaze1518_sinhnhatf168_internal_errors.Meta:
    properties:
      links:
        $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.Links'
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      timestamp:
        type: string
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_Blaze1518_sinhnhatf168_internal_errors.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.Meta'
      success:
        type: boolean
    type: object
  internal_game_campaign.Campaign:
    properties:
      code:
        type: string
      created_at:
        type: string
      ended_at:
        type: string
      id:
        type: integer
      name:
        type: string
      started_at:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  internal_game_campaign.CreateCampaignRequest:
    properties:
      code:
        example: SINHNHAT_F168_2026
        maxLength: 100
        type: string
      ended_at:
        example: "2026-06-26T23:59:59Z"
        type: string
      name:
        example: Chương trình Khuyến mãi Sinh nhật F168
        maxLength: 255
        type: string
      started_at:
        example: "2026-05-26T00:00:00Z"
        type: string
      status:
        enum:
        - DRAFT
        - ACTIVE
        - PAUSED
        - ENDED
        example: ACTIVE
        type: string
    required:
    - code
    - name
    type: object
  internal_game_prize.CreatePrizeRequest:
    properties:
      campaign_id:
        example: 1
        type: integer
      name:
        example: Xe máy Honda Vision 2026
        maxLength: 255
        type: string
      priority:
        example: 1
        type: integer
      quantity:
        example: 10
        minimum: 1
        type: integer
    required:
    - campaign_id
    - name
    - quantity
    type: object
  internal_game_prize.Prize:
    properties:
      campaign_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      priority:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  internal_game_ticket.CreateTicketRequest:
    properties:
      campaign_id:
        example: 1
        type: integer
    required:
    - campaign_id
    type: object
  internal_game_ticket.Ticket:
    properties:
      campaign_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_canceled:
        type: boolean
      is_winner:
        type: boolean
      ticket_number:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  internal_game_ticket.TicketImportJob:
    properties:
      campaignID:
        type: integer
      createdAt:
        type: string
      errorLog:
        type: string
      failedRows:
        type: integer
      id:
        type: integer
      status:
        type: string
      successRows:
        type: integer
      totalRows:
        type: integer
      updatedAt:
        type: string
    type: object
  internal_game_winner.DrawRequest:
    properties:
      campaign_id:
        example: 1
        type: integer
      prize_id:
        example: 2
        type: integer
    required:
    - campaign_id
    - prize_id
    type: object
  internal_game_winner.Winner:
    properties:
      campaign_id:
        type: integer
      created_at:
        type: string
      draw_order:
        type: integer
      id:
        type: integer
      prize_id:
        type: integer
      ticket_id:
        type: integer
    type: object
  whitelistip.CreateWhitelistIPRequest:
    properties:
      description:
        description: Description optional human-readable note (e.g. office name, ticket
          id).
        example: HQ VPN egress - INC-1234
        maxLength: 255
        type: string
      ip_address:
        example: 203.0.113.10/32
        maxLength: 64
        type: string
      is_active:
        description: IsActive when false, the entry is persisted but must not grant
          access (enforcement is middleware responsibility).
        example: true
        type: boolean
    required:
    - ip_address
    - is_active
    type: object
host: localhost:8081
info:
  contact: {}
  description: API documentation for HappyBirthDayF168
  termsOfService: http://swagger.io/terms/
  title: HappyBirthDayF168
  version: "1.0"
paths:
  /game/campaign:
    post:
      consumes:
      - application/json
      description: Tạo mới một chiến dịch với tên và mã code riêng biệt. Hệ thống
        tự động chặn trùng mã code.
      parameters:
      - description: Thông tin chiến dịch cần tạo mới
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_game_campaign.CreateCampaignRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Tạo chiến dịch thành công
          schema:
            $ref: '#/definitions/internal_game_campaign.Campaign'
        "400":
          description: Dữ liệu gửi lên không hợp lệ hoặc sai định dạng JSON
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.APIError'
        "409":
          description: Mã định danh chiến dịch (Code) đã tồn tại
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.APIError'
        "500":
          description: Lỗi hệ thống nội bộ
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.APIError'
      summary: Tạo mới một chiến dịch (Campaign)
      tags:
      - Campaigns
  /game/prize:
    post:
      consumes:
      - application/json
      description: Tạo mới một giải thưởng thuộc về một chiến dịch cụ thể bằng campaign_id.
      parameters:
      - description: Thông tin giải thưởng cần tạo mới
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_game_prize.CreatePrizeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Tạo giải thưởng thành công
          schema:
            $ref: '#/definitions/internal_game_prize.Prize'
        "400":
          description: Dữ liệu gửi lên không hợp lệ hoặc sai định dạng JSON
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.APIError'
        "500":
          description: Lỗi hệ thống nội bộ
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.APIError'
      summary: Tạo mới một giải thưởng (Prize)
      tags:
      - Prizes
  /game/ticket:
    post:
      consumes:
      - application/json
      description: Tạo mới một lượt chơi/vé số thuộc về một chiến dịch. Mã vé dài
        21 ký tự sẽ tự động được sinh ngẫu nhiên bảo mật ở hệ thống.
      parameters:
      - description: Thông tin chiến dịch để cấp vé
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_game_ticket.CreateTicketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Phát hành vé thành công
          schema:
            $ref: '#/definitions/internal_game_ticket.Ticket'
        "400":
          description: Dữ liệu gửi lên không hợp lệ hoặc sai định dạng JSON
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.APIError'
        "500":
          description: Lỗi hệ thống nội bộ
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.APIError'
      summary: Phát hành một vé số mới (Ticket)
      tags:
      - Tickets
  /game/ticket/import:
    post:
      consumes:
      - multipart/form-data
      description: Nhận file .xlsx xử lý bất đồng bộ dưới nền.
      parameters:
      - description: ID của chiến dịch
        in: formData
        name: campaign_id
        required: true
        type: integer
      - description: File Excel chứa danh sách mã vé (.xlsx)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Đã tiếp nhận file và đang xử lý ngầm
          schema:
            $ref: '#/definitions/internal_game_ticket.TicketImportJob'
      summary: Upload file Excel để import lượng lớn vé số
      tags:
      - Tickets
  /game/winner/draw:
    post:
      consumes:
      - application/json
      description: Thực hiện bốc thăm ngẫu nhiên bảo mật dưới dạng Transaction để
        tìm vé trúng thưởng, tự động trừ kho giải và hủy các vé còn lại của user trúng
        giải.
      parameters:
      - description: Thông tin chiến dịch và giải thưởng cần quay
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_game_winner.DrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Quay số thành công và đã xác định người trúng giải
          schema:
            $ref: '#/definitions/internal_game_winner.Winner'
        "400":
          description: Lỗi nghiệp vụ (Hết giải thưởng, Hết vé hợp lệ hoặc sai dữ liệu
            đầu vào)
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.APIError'
        "500":
          description: Lỗi hệ thống nội bộ
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.APIError'
      summary: Kích hoạt quay số xác định người trúng giải (Draw)
      tags:
      - Winners
  /whitelist-ips/:
    post:
      consumes:
      - application/json
      description: |-
        Creates a unique IP or CIDR row in table whitelist_ips. Stored as PostgreSQL cidr: use a network prefix (e.g. 203.0.113.10/32 for a single IPv4 host, 2001:db8::/32 for an IPv6 range). Duplicate ip_address returns 409. is_active is stored as-is; access control must ignore rows with is_active=false.
        On success the JSON envelope uses success/data; data is a WhitelistIP object (id, ip_address, description, is_active, created_at, updated_at).
      operationId: createWhitelistIP
      parameters:
      - description: Create whitelist IP payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/whitelistip.CreateWhitelistIPRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created; data field contains WhitelistIP
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.Response'
        "400":
          description: Validation error (invalid JSON or binding rules)
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.Response'
        "409":
          description: Conflict — IP/CIDR already exists
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.Response'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/github_com_Blaze1518_sinhnhatf168_internal_errors.Response'
      summary: Create whitelist IP
      tags:
      - whitelist-ips
swagger: "2.0"